特征提取切换菜单时防止重新渲染播放器
